+++ .gitignore
@@ -1,0 +1,8 @@
+﻿################################################################################
+# This .gitignore file was automatically created by Microsoft(R) Visual Studio.
+################################################################################
+
+/.vs
+/obj
+/com.dreadscripts.vrcsdkplus.csproj
+/com.dreadscripts.vrcsdkplus.sln
