+++ Editor/EventCommands.cs
@@ -1,0 +1,17 @@
+﻿namespace DreadScripts.VRCSDKPlus
+{
+    internal enum EventCommands
+    {
+        Copy,
+        Cut,
+        Paste,
+        Duplicate,
+        Delete,
+        SoftDelete,
+        SelectAll,
+        Find,
+        FrameSelected,
+        FrameSelectedWithLock,
+        FocusProjectWindow
+    }
+}
