+++ Editor/PathOption.cs
@@ -1,0 +1,9 @@
+﻿namespace DreadScripts.VRCSDKPlus
+{
+    internal enum PathOption
+    {
+        Normal,
+        ForceFolder,
+        ForceFile
+    }
+}
