+++ Editor/Preferences.cs
@@ -1,0 +1,20 @@
+﻿using UnityEditor;
+
+namespace DreadScripts.VRCSDKPlus
+{
+    public static class Preferences
+    {
+        public static bool CompactMode
+        {
+            get
+            {
+                return EditorPrefs.GetBool(Strings.SettingsCompact, false);
+            }
+
+            set
+            {
+                EditorPrefs.SetBool(Strings.SettingsCompact, value);
+            }
+        }
+    }
+}
