+++ Editor/Strings.cs
@@ -1,0 +1,22 @@
+﻿namespace DreadScripts.VRCSDKPlus
+{
+    public static class Strings
+    {
+        public const string IconCopy = "SaveActive";
+        public const string IconPaste = "Clipboard";
+        public const string IconMove = "MoveTool";
+        public const string IconPlace = "DefaultSorting";
+        public const string IconDuplicate = "TreeEditor.Duplicate";
+        public const string IconHelp = "_Help";
+        public const string IconWarn = "console.warnicon.sml";
+        public const string IconError = "console.erroricon.sml";
+        public const string IconClear = "winbtn_win_close";
+        public const string IconFolder = "FolderOpened Icon";
+        public const string IconRemove = "Toolbar Minus";
+        public const string IconSearch = "Search Icon";
+
+        public const string ClipboardPrefixControl = "[TAG=VSP_CONTROL]";
+
+        public const string SettingsCompact = "VSP_Compact";
+    }
+}
